requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
pymysql==1.0.4
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.12
